Implemented battle in UI

+++ Monster_Hunter/Domain/Game/Character.hpp
@@ -8,13 +8,13 @@ namespace Domain::Game
 	class Character
 	{
 	protected:
-		int attributes;
+		double attributes;
 		double max_health;
 		double current_health;
 		int potion;
-		int base_damage;
-		int damage;
-		int weapon_attributes;
+		double base_damage;
+		double damage;
+		double weapon_attributes;
 		std::string name;
 		std::array<std::array<std::string, 20>, 20> image;
 		std::string weapon_name;
@@ -26,23 +26,23 @@ namespace Domain::Game
 		Character(Character && original) = default;			//Move Constructor can be removed later
 		//~Character() = default;					//Virtual Destructor
 		virtual ~Character() noexcept = 0;
-		void gain_Attribute(int value, int value_health);
-		void equip_Weapon(int wa, std::string wn);
+		void gain_Attribute(double value, double value_health);
+		void equip_Weapon(double wa, std::string wn);
 		virtual void calculate_damage() = 0;
-		int get_attributes();
-		void set_attributes(const int value);
+		double get_attributes();
+		void set_attributes(const double value);
 		double get_max_health();
 		void set_max_health(const double value);
 		double get_current_health();
 		void set_current_health(const double value);
 		int get_potion();
 		void set_potion(const int value);
-		int get_base_damage();
-		void set_base_damage(const int value);
-		int get_damage();
-		void set_damage(const int value);
-		int get_weapon_attributes();
-		void set_weapon_attributes(const int value);
+		double get_base_damage();
+		void set_base_damage(const double value);
+		double get_damage();
+		void set_damage(const double value);
+		double get_weapon_attributes();
+		void set_weapon_attributes(const double value);
 		std::string get_weapon_name();
 		void set_weapon_name(const std::string value);
 		std::string get_name();
